Don't allow deleting required questions (#701)

+++ limber-backend-application/module/forms/forms-application/src/main/kotlin/io/limberapp/backend/module/forms/service/formInstance/FormInstanceQuestionServiceImpl.kt
@@ -2,12 +2,17 @@ package io.limberapp.backend.module.forms.service.formInstance
 
 import com.google.inject.Inject
 import com.piperframework.finder.Finder
+import io.limberapp.backend.module.forms.exception.formInstance.CannotDeleteRequiredQuestion
+import io.limberapp.backend.module.forms.exception.formInstance.FormInstanceQuestionNotFound
 import io.limberapp.backend.module.forms.model.formInstance.FormInstanceQuestionFinder
 import io.limberapp.backend.module.forms.model.formInstance.FormInstanceQuestionModel
+import io.limberapp.backend.module.forms.service.formTemplate.FormTemplateQuestionService
 import io.limberapp.backend.module.forms.store.formInstance.FormInstanceQuestionStore
 import java.util.*
 
 internal class FormInstanceQuestionServiceImpl @Inject constructor(
+  private val formTemplateQuestionService: FormTemplateQuestionService,
+  private val formInstanceService: FormInstanceService,
   private val formInstanceQuestionStore: FormInstanceQuestionStore,
 ) : FormInstanceQuestionService,
   Finder<FormInstanceQuestionModel, FormInstanceQuestionFinder> by formInstanceQuestionStore {
@@ -16,6 +21,17 @@ internal class FormInstanceQuestionServiceImpl @Inject constructor(
     return formInstanceQuestionStore.upsert(featureGuid, model)
   }
 
-  override fun delete(featureGuid: UUID, formInstanceGuid: UUID, questionGuid: UUID) =
+  override fun delete(featureGuid: UUID, formInstanceGuid: UUID, questionGuid: UUID) {
+    val formInstance = formInstanceService.findOnlyOrNull {
+      featureGuid(featureGuid)
+      formInstanceGuid(formInstanceGuid)
+    } ?: throw FormInstanceQuestionNotFound()
+    val formTemplateQuestion = formTemplateQuestionService.findOnlyOrNull {
+      featureGuid(featureGuid)
+      formTemplateGuid(formInstance.formTemplateGuid)
+      questionGuid(questionGuid)
+    } ?: throw FormInstanceQuestionNotFound()
+    if (formTemplateQuestion.required && formInstance.submittedDate != null) throw CannotDeleteRequiredQuestion()
     formInstanceQuestionStore.delete(featureGuid, formInstanceGuid, questionGuid)
+  }
 }
